refactor(code-generation): improve type safety and clean up file processing logic
