+++ generate/code_generation.ts
@@ -1,0 +1,90 @@
+import type {Config} from "./config";
+import {writeFile} from "fs/promises";
+import {fileIcons} from "./vscode-icons/src/defaults/fileIcons";
+
+const template = `// This file is automatically generated by Assets Builder.
+// DO NOT MODIFY IT MANUALLY!
+
+package cat.wavy.catactivity.types
+
+enum class FileType(
+    val typeName: String,
+    val icon: String
+) {
+    %fileTypes%
+}
+
+fun getFileTypeByName(fileType: String, fileName: String, extension: String?) = when (fileName.lowercase()) {
+    %fileNameMatch%
+    else -> when (fileType) {
+        %fileTypeMatch%
+        else -> when (extension?.lowercase()) {
+            %fileExtensionMatch%
+            else -> FileType.FILE
+        }
+    }
+}`
+
+export type Files = {
+  enumName: string
+  name: string
+  icon: string
+  fileNames: string[] | null | undefined
+  fileTypes: string[] | null | undefined
+  fileExtensions: string[] | null | undefined
+}
+
+type FileListProperty = {
+  [K in keyof Files]: Files[K] extends string[] | null | undefined ? K : never
+}[keyof Files]
+
+function formatFileTypes(files: Files[]): string {
+  return files
+    .map(file => `${file.enumName}("${file.name}", "${file.icon}")`)
+    .join(",\n    ")
+}
+
+function formatFilesMatch(files: Files[], propertyName: FileListProperty, tabNum: number): string {
+  const indent = '\n' + ' '.repeat(4 * tabNum)
+  const items: string[] = []
+
+  for (const file of files) {
+    const values = file[propertyName]
+    if (!values || values.length === 0) continue
+    items.push(`"${values.join('", "')}" -> FileType.${file.enumName}`)
+  }
+
+  return items.join(indent)
+}
+
+export async function code_generation(config: Config) {
+  const items: Files[] = []
+
+  for (const icon in config.fileIcons) {
+    const property = config.fileIcons[icon]
+    if (!property) continue
+
+    let enumName = property.enumName ? property.enumName : icon.toUpperCase().replace(/-/g, "_")
+    if (/^\d/.test(enumName)) enumName = "A_" + enumName
+
+    const defaults = fileIcons[icon as keyof typeof fileIcons]
+
+    items.push({
+      icon: icon,
+      name: property.title,
+      enumName: enumName,
+      fileTypes: property.fileTypes,
+      fileExtensions: property.extensions === undefined ? defaults?.fileExtensions : property.extensions,
+      fileNames: property.fileNames === undefined ? defaults?.fileNames : property.fileNames
+    })
+  }
+
+  const code = template
+    .replace('%fileTypes%', formatFileTypes(items))
+    .replace('%fileNameMatch%', formatFilesMatch(items, 'fileNames', 1))
+    .replace('%fileTypeMatch%', formatFilesMatch(items, 'fileTypes', 2))
+    .replace('%fileExtensionMatch%', formatFilesMatch(items, 'fileExtensions', 3))
+  const outputPath = 'src/main/kotlin/cat/wavy/catactivity/types/FileType.kt'
+
+  await writeFile(outputPath, code)
+}
